+++ src/us/mn/state/dot/tms/client/system/PrivilegePanel.java
@@ -1,6 +1,6 @@
 /*
  * IRIS -- Intelligent Roadway Information System
- * Copyright (C) 2016-2021  Minnesota Department of Transportation
+ * Copyright (C) 2016-2022  Minnesota Department of Transportation
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -39,7 +39,6 @@ public class PrivilegePanel extends ProxyTablePanel<Privilege> {
 		"alert_message",
 		"beacon",
 		"beacon_action",
-		"cabinet",
 		"cabinet_style",
 		"camera",
 		"camera_action",